+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- compatibility with python3.5:
+  - fixed `TypeError` due to `start_engine` passing `pathlib.Path`
+    to `subprocess.Popen` for `cwd`
+  - tests: fixed `AttributeError` due to unavailable `MagicMock.assert_called_once`
+  - tests: fixed `FileNotFoundError` due to strict `pathlib.Path.resolve`
 
 ## [1.0.2] - 2020-05-05
 ### Fixed